pyproject.toml
[tool.poetry]
name = "py_site_generator"
version = "0.1.0"
description = ""
authors = ["Ilya Siganov <ilya.blan4@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Markdown = "^3.3.7"
pymdown-extensions = "^9.5"
tqdm = "^4.64.0"
python-frontmatter = "^1.0.0"
Jinja2 = "^3.1.2"
libsass = "^0.21.0"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.16.1"
autopep8 = "^1.6.0"
isort = "^5.10.1"
pydantic = "^1.9.1"

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools"]
build-backend = "poetry.core.masonry.api"

[project]
name = "ssg"
description = "SSG: simple static site generator."
readme = "README.md"
license = {text= "MIT"}
classifiers=[
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3 :: Only',
    'Environment :: Console',
    'Topic :: Utilities'
]

[project.scripts]
ssg = "ssg.__main__:main"